reorder fixure loading
